Fix detection of Mac OS X.
reference to Mono.Posix.dll must be added to the project. The file must be copied to the Mono-D distribution (./ext/Mono.Posix.dll), because it is not present on the Windows distribution of Xamarin Studio.
Mono.Posix.dll can be found here:
/Applications/Xamarin Studio.app/Contents/Resources/lib/monodevelop/AddIns/Xamarin.Sketches/XamarinInteractiveAgentsMac.app/Contents/MonoBundle/Mono.Posix.dll

The initial check for Unix will prevent the posix code from trying to run on windows.